Refactor layout file and use ref
